Fix overflow

+++ tests/vdp/ports.cpp
@@ -620,8 +620,8 @@ TEST(VDP_PORTS, DATA_PORT_READ_VRAM_AUTO_INC)
 	auto& mem = vdp.vram();
 
 	// prepare mem
-	for(std::uint16_t addr = 0; addr <= mem.max_address; ++addr)
-		mem.write(addr, test::random::next<std::uint8_t>());
+	for(std::size_t addr = 0; addr <= mem.max_address; ++addr)
+		mem.write(static_cast<std::uint16_t>(addr), test::random::next<std::uint8_t>());
 
 	control_register control = setup_control(0, vmem_type::vram, control_type::read);
